implement policies for the rest of the models except user

+++ app/Http/Controllers/GameController.php
@@ -40,6 +40,10 @@ class GameController extends Controller
 
         $game = Game::where('name', $game_name['game_name'])->firstOrFail();
 
+        if ($request->user()->cannot('store', [Game::class, $request->user()])) {
+            abort(403);
+        }
+
         if (Session::has('impersonate')) {
             $id = Session::get('impersonate');
         } else {
@@ -47,7 +51,7 @@ class GameController extends Controller
         }
 
         DB::table('gamemasters')->insert([
-            'id' => $id,
+            'user_id' => $id,
             'game_id' => $game->id,
         ]);
 
@@ -65,7 +69,7 @@ class GameController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $game_id)
+    public function edit(Request $request, string $game_id)
     {
         $id = Auth::user()->id;
         if (Session::has('impersonate')) {
@@ -95,6 +99,10 @@ class GameController extends Controller
 
         $game = Game::find($game_id);
 
+        if ($request->user()->cannot('update', $game)) {
+            abort(403);
+        }
+
         $game->name = $validated['game_name'];
         $game->save();
         $game->update();
@@ -105,9 +113,14 @@ class GameController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, string $id)
     {
         $game = Game::findOrFail($id);
+
+        if ($request->user()->cannot('delete', $game)) {
+            abort(403);
+        }
+
         $game->delete();
 
         return redirect()->route('game.index');
